JDWP: implement frame variable access, thread suspend tracking, and method invocation fix
Frame variable access (locals/dump/set):
- Add framePointer field to VmStackFrame, propagated from VmStackReader
  through VmStackFrameEnumerator.getFramePtr()
- Implement readLocalVariable(slot) and writeLocalVariable(slot) in
  VmStackFrame using EBP-based offset calculation matching the
  X86StackFrame.getEbpOffset() formula from all three compilers
  (l1a/l1b/l2): args at (argSlotCount - slot + 2) * slotSize above
  EBP, locals at (slot - argSlotCount + 1) * -slotSize below EBP
- Type-aware read/write via LocalVariableTable descriptor lookup
  (I/J/F/D/B/S/C/Z for primitives, object reference fallback)
- NativeVMFrame.getValue/setValue delegate to VmStackFrame
  readLocalVariable/writeLocalVariable; primitives supported, object
  reference writes deferred (require GC integration)
- VMFrame gets vmStackFrame field (package-private) linking JDWP frame
  to native VmStackFrame for variable access
- buildFrame() reads slot 0 ('this') for instance methods via
  readLocalVariable(0)

Thread suspend/resume tracking:
- JDIVirtualMachine: HashMap<Thread,Integer> suspendCounts tracks
  per-thread suspend counts (increment/decrement semantics)
- suspendThread/resumeThread/getSuspendCount implement actual tracking
  instead of no-ops
- ThreadReferenceCommandSet.executeStatus() reports SUSPENDED vs
  NOT_SUSPENDED based on getSuspendCount() instead of hardcoded
  NOT_SUSPENDED

Null safety and protocol fixes:
- Value.writeValue(): handle null objects (write OBJECT tag + null ID)
  preventing NPE that caused JDWP INTERNAL (113) errors
- StackFrameCommandSet: null frame guards in executeGetValues,
  executeSetValues, executeThisObject (return empty/null responses)
- VariableTable.write(): fix argCnt/slots written as writeLong (8
  bytes) to writeInt (4 bytes) per JDWP spec (caused
  StringIndexOutOfBoundsException in jdb)

Method invocation fix (print/dump):
- ObjectReferenceCommandSet.executeInvokeMethod and
  ClassTypeCommandSet.invokeMethod: read method ID as raw long index
  and resolve VmMethod from VmType.getDeclaredMethod(idx) instead of
  idMan.readObjectId() which collided with object ID space (method index
  1 mapped to IRQThread object at object ID 1)

+++ core/src/classpath/ext/gnu/classpath/jdwp/util/VariableTable.java
@@ -95,9 +95,9 @@ public class VariableTable
    */
   public void write(DataOutputStream os) throws IOException
   {
-    os.writeLong(argCnt);
-    os.writeLong(slots);
-    for (int i = 0; i < slots; i++)
+    os.writeInt((int) argCnt);
+    os.writeInt((int) slots);
+    for (int i = 0; i < (int) slots; i++)
       {
         os.writeLong(lineCI[i]);
         JdwpString.writeString(os, names[i]);
